refactor: move cross-layer constants to shared/constants.py
