+++ src/quodeq/shared/constants.py
@@ -1,0 +1,8 @@
+"""Cross-layer constants shared across bounded contexts."""
+
+# Pool / subagent defaults
+_DEFAULT_MAX_SUBAGENTS = 5
+_DEFAULT_POOL_BUDGET = 600  # 10 minutes total pool budget (seconds)
+
+# Structured marker key used for job-tracking JSON markers
+CC_MARKER_KEY = "_cc"
